# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Web OTP API",
    "version": "1.0.0",
    "description": "Issue and verify email one-time passwords. OTPs are 6-digit codes valid for 1 hour."
  },
  "servers": [
    { "url": "/api", "description": "Production" }
  ],
  "paths": {
    "/otp": {
      "post": {
        "summary": "Issue OTP",
        "description": "Generate a 6-digit OTP for the given email address and send it via Gmail.",
        "operationId": "issueOTP",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/IssueRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" },
                "example": { "message": "OTP sent to your email" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/otp/verify": {
      "post": {
        "summary": "Verify OTP",
        "description": "Check whether a given OTP is valid for the email. If valid, the OTP is consumed (cannot be reused).",
        "operationId": "verifyOTP",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerifyRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VerifyResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/stats": {
      "get": {
        "summary": "Usage statistics",
        "description": "Returns summary counts and a paginated list of OTP records. Supports filtering by date range and email.",
        "operationId": "getStats",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "description": "Include records created on or after this datetime (ISO 8601)",
            "schema": { "type": "string", "format": "date-time" },
            "example": "2024-01-01T00:00:00Z"
          },
          {
            "name": "to",
            "in": "query",
            "description": "Include records created on or before this datetime (ISO 8601)",
            "schema": { "type": "string", "format": "date-time" },
            "example": "2024-12-31T23:59:59Z"
          },
          {
            "name": "email",
            "in": "query",
            "description": "Partial match filter on email address",
            "schema": { "type": "string" },
            "example": "user@"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of records to return (max 100)",
            "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of records to skip for pagination",
            "schema": { "type": "integer", "default": 0, "minimum": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "Stats and paginated records",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StatsResponse" }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "IssueRequest": {
        "type": "object",
        "required": ["email", "session_token"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "session_token": {
            "type": "string",
            "minLength": 8,
            "description": "Random token generated client-side (e.g. crypto.randomUUID()). Must be re-sent on verify to bind the OTP to this browser session.",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        }
      },
      "VerifyRequest": {
        "type": "object",
        "required": ["email", "otp", "session_token"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "otp": {
            "type": "string",
            "pattern": "^[0-9]{6}$",
            "example": "482910"
          },
          "session_token": {
            "type": "string",
            "description": "The same token sent when issuing the OTP.",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        }
      },
      "VerifyResponse": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean",
            "description": "true if the OTP was correct and not expired; false otherwise"
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        }
      },
      "OTPRecord": {
        "type": "object",
        "properties": {
          "id":         { "type": "integer", "format": "int64" },
          "email":      { "type": "string", "format": "email" },
          "created_at": { "type": "string", "format": "date-time" },
          "expires_at": { "type": "string", "format": "date-time" },
          "used":       { "type": "boolean" },
          "used_at":    { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "StatsResponse": {
        "type": "object",
        "properties": {
          "summary": {
            "type": "object",
            "properties": {
              "total_issued":   { "type": "integer" },
              "total_verified": { "type": "integer" },
              "total_expired":  { "type": "integer" },
              "total_pending":  { "type": "integer" }
            }
          },
          "records": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/OTPRecord" }
          },
          "total": {
            "type": "integer",
            "description": "Total matching records (before pagination)"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request body or missing field",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "valid email is required" }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "failed to issue OTP" }
          }
        }
      }
    }
  }
}